+++ src/typing_ex/view/page.clj
@@ -11,7 +11,7 @@
    [typing-ex.plot :refer [scatter]]
    #_[clojure.test :as t]))
 
-(def ^:private version "2.2.803")
+(def ^:private version "v2.5.836")
 
 ;--------------------------------
 ;; FIXME
@@ -40,9 +40,8 @@
      [:meta {:name "viewport" :content "width=device-width, initial-scale=1"}]]
     [:link
      {:rel "stylesheet"
-      :href "https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css"
-      :integrity "sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65"
-      :crossorigin "anonymous"}]
+      :href "/css/bootstrap.min.css"
+      :type "text/css"}]
     [:link
      {:rel "stylesheet"
       :href "/css/style.css"}]
@@ -78,10 +77,10 @@
      [:a {:href "/" :class "btn btn-primary btn-sm"} "Go!"]
      "&nbsp;"
      ;;py99
-     [:a {:href "https://py99.melt.kyutech.ac.jp/"
-          :class "btn btn-info btn-sm"}
-      "Py99"]
-     "&nbsp;"
+    ;;  [:a {:href "https://py99.melt.kyutech.ac.jp/"
+    ;;       :class "btn btn-info btn-sm"}
+    ;;   "Py99"]
+    ;;  "&nbsp;"
      ;; literacy reports
      ;;  [:a {:href "https://rp.melt.kyutech.ac.jp/"
      ;;       :class "btn btn-info btn-sm"}
@@ -90,7 +89,7 @@
      [:a {:href "/rc" :class "btn btn-info btn-sm"} "RC"]
      "&nbsp;"
      [:a {:href "https://wil.melt.kyutech.ac.jp/"
-          :class "btn btn-info btn-sm"}
+          :class "btn btn-success btn-sm"}
       "WIL"]
      "&nbsp;"
      [:a {:href "http://qa.melt.kyutech.ac.jp/"
@@ -102,7 +101,7 @@
       "MT"]
      "&nbsp;"
      [:a {:href "http://l22.melt.kyutech.ac.jp/"
-          :class "btn btn-info btn-sm"}
+          :class "btn btn-success btn-sm"}
       "L22"]
      "&nbsp;"
      [:a {:href "/logout" :class "btn btn-warning btn-sm"} "Logout"]]]
@@ -113,6 +112,10 @@
      "&nbsp;"
      (form-to
       [:get "/recent"]
+      (submit-button {:class "btn btn-primary btn-sm"
+                      :name "kind"}
+                     "training days")
+      "&nbsp;"
       (text-field {:size 2
                    :value n
                    :style "text-align:right"}
@@ -121,10 +124,6 @@
       (submit-button {:class "btn btn-primary btn-sm"
                       :name "kind"}
                      "total")
-      "&nbsp;"
-      (submit-button {:class "btn btn-primary btn-sm"
-                      :name "kind"}
-                     "days")
       "&nbsp;"
       (submit-button {:class "btn btn-primary btn-sm"
                       :name "kind"}
@@ -162,30 +161,49 @@
        count))
 
 (defn ex-days-page
-  "ex-days: 練習日数
-   user: アカウント
-   days: 何日間のデータか？"
-  [ex-days user days]
-  (let [logins (->> ex-days (map :login) distinct)
-        data (->> (for [login logins]
-                    [(count-ex-days ex-days login) login])
-                  (sort-by first)
-                  reverse)]
-    (page
-     [:h2 "Typing: Last " days " days Maxes"]
-     (headline days)
-     [:div {:style "margin-left:1rem;"}
-      [:p "毎日ちょっとずつが伸びる秘訣。"]
-      (into [:ol
-             (for [[count login] data]
-               [:li
-                (format "(%d) " count)
-                " "
-                [:a {:href (str "/record/" login)
-                     :class (if (= login user) "yes" "other")}
-                 login]])])]
-     ;; (headline days)
-     )))
+  "self はログインアカウント、
+   data はソーティング済みの[[login days] ...]"
+  [self data]
+  (page
+   [:h2 "Typing: 10 回以上練習した日数"]
+   (headline 7)
+   [:div {:style "margin-left:1rem;"}
+    [:p "毎日ちょっとずつが伸びる秘訣。"]
+    (into [:ol
+           (for [[login n] data]
+             [:li
+              (format "(%d) " n)
+              " "
+              [:a {:href (str "/record/" login)
+                   :class (if (= login self) "yes" "other")}
+               login]])])]))
+
+;; (defn ex-days-page
+;;   "ex-days: 練習日数
+;;    user: アカウント
+;;    days: 何日間のデータか？"
+;;   [ex-days user days]
+;;   (let [logins (->> ex-days (map :login) distinct)
+;;         data (->> (for [login logins]
+;;                     [(count-ex-days ex-days login) login])
+;;                   (sort-by first)
+;;                   reverse)]
+;;     (page
+;;      [:h2 "Typing: Last " days " days Maxes"]
+;;      (headline days)
+;;      [:div {:style "margin-left:1rem;"}
+;;       [:p "毎日ちょっとずつが伸びる秘訣。"]
+;;       (into [:ol
+;;              (for [[count login] data]
+;;                [:li
+;;                 (format "(%d) " count)
+;;                 " "
+;;                 [:a {:href (str "/record/" login)
+;;                      :class (if (= login user) "yes" "other")}
+;;                  login]])])]
+;;      ;; (headline days)
+;;      )))
+
 
 (defn- select-count-distinct
   "select count(distinct(timestamp::DATE)) from results